Refactor Tests for .FOUR and FOUR Measure Output
This commit adds a comparator (compare_fourier_files.pl) that is
specific to the output files produced by .FOUR lines and FOUR
measures.  It allows for a different tolerance (phaseAbsTol) for
the phase values.  It is also aware of the specific fields in the
header lines that are integer-valued.

That comparator is then used by the relevant .sh files in FOURIER,
Output/Dasho and Certification_Tests/BUG_1962.  This replaces a
lot of repeated Perl code in those shell files, while improving
the testing of the output formats for .FOUR and FOUR measures.

This commit also renames some of the "gold standards" to have .four0
or .mt0 file extensions.  This is more consistent with the rest of
testing.  It also allows the formatting/contents of the slightly
different header lines, in the .FOUR and FOUR measure output files,
to be verified.

The test FOURIER/lead_current_not_on_print_line_four was simplified.
It was not actually testing "lead currents not on the .PRINT TRAN
or .MEASURE TRAN lines", and those subtests are redundant with
other tests in the test suite anyway.  So, those subtests were
deleted from the netlist and the test's .sh file.

The final change is that some .sh files ran their test twice;
once with a netlist with a .FOUR line and a second time with the
corresponding .MEASURE line.  That is unnecessary now, and those
subtests have been combined into one netlist.

Issues #222 and #248
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/222
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/248