Refactor Tests for FOUR Measures
This commit refactors the tests for the FOUR measures in MEASURE
and MEASURE/STEP to make them simpler, and more similar to the
rest of the measure tests.  It also adds testing of the stdout
for FOUR measures for both the .STEP and non-step cases.

The comparator compare_fourier_files.pl was modified to allow a
separate phaseRelTol for the phase fields in the .FOUR and FOUR
measure output.  This was most helpful for the MEASURE/FourierTest
test.  This change required an update to the other .sh files for
the tests of .FOUR and FOUR measure output in FOURIER, Output/Dasho
and Certification_Tests/BUG_1962.

This commit also improves the subtests for the number format and
precision in the FOUR measure output.  The integer fields are
now verified to be integers.  In addition, the number format
and precision in both the .mt0 file and the stdout are now
verified.  It also tests .OPTIONS MEASURE MEASDGT and the
PRECISION qualifier on FOUR measures.

Issue #248 and #226
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/248
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/226