Add a test case for issue 243/GH29
This commit adds a test to check that a program that uses the general
external output interface will correctly output "I(*)" wildcards
irrespective of whether the netlist it is running has a .print line
with I(*) on it.  Prior to the fix of issue 243, it wouldn't --- I(*)
would not be delivered to the external output interfaces unless some
.print line in the netlist  had it.

The test builds a simple test program that creates an external
outputter, and takes the same command line arguments as Xyce itself.
The outputter is one that tries to imitate .PRN formatting
and outputs to a file called "ioTest.out".  Only transient is
currently supported in this toy program (though it could easily be
adapted to other analysis types). The external outputter requests
"V(*) I(*)".

This small toy program is then run over two netlists --- one with no
print line at all, and another with a `.print tran V(*) I(*)` line.

The ioTest.out files produced by the two runs should be identical, and
are diffed.  Since they are both produced by the same program on the
same platform, using diff is OK and xplat_diff is not needed.

If the files produced by the external outputter on the two netlists
are identical, then xyce_verify is run to compare the .print output to
the external output.  Because xyce_verify does not understand
wildcards in print lines, the header of the .print output is scanned
and a fake netlist with a .print line that would produce that output
set is created.  This fake netlist is fed to xyce_verify to tell it
what header to expect, and the .print output is compared to the
external IO output.

If this test were run with Xyce 7.3, it would fail because the two
ioTest.out files would not be identical --- the one produced by the
run over the netlist without a print line would be missing current
columns.

The test program is called "TestIssue29" and the test is called
TestGH29 because this issue was first detected while servicing a
github issue (issue 28, which turned out to be user error), and a
github issue 29 was created.  But the gitlab-ex issue 243 was opened
to track its progress internally.

Xyce#243
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/243
https://github.com/Xyce/Xyce/issues/29