Test Use of FOUR Measures on .PRINT TRAN Lines
This commit adds FOUR measures to the .PRINT TRAN lines in two
existing tests in MEASURE.  It verifies that the output to stdout
and the <netlistName>.mt0 file is still correct for both tests.

The measure values in the <netlistName>.prn files are incorrectly
reported as always -1, per the comments in the associated commit
for this issue to the Xyce repo.

Issue #255
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/255