Fix source parameters in HSPICE netlists
The HSPICE netlists i_sources_exp.cir.hspice and i_sources_sffm.cir.hspice
don't have the same meaning as the equivalvent Xyce netlist. The
"A" used in the source parameters for Xyce means amps, while in
HSPICE (and in Xyce when using -hspice-ext all) it means atto. So
this was fixed by removing the "A" suffix.