Test Use of I(*) and P(*) with LTRA Device in Netlist
The previous version of the function Xyce::IO::removeWildcardVariables()
had special handling for the O device, which does not support either
lead currents or power.  This commit adds a regression test that
verifies that "special handling" is no longer needed.  The O device
is safely ignored by both I(*) and P(*) wildcards.

This test will fail, because of the I1(*) and I2(*) wildcards, if/when
lead currents are implemented for the O device.  The fix to the .sh
file should then be a simple change to the @expectedOutputs array.

Issue #254
SON Bugs 983 and 1317
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=983
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1317