Adds test to check correct .PREPROCESS commands are commented out
If the .PREPROCESS ADDRESISTORS command is used, Xyce will generate a
new netlist that includes a copy of the commented out .PREPROCESS command
to serve as a reminder that the netlist resulted from this preprocessing step.
Unfortunately, the logic to do this was flawed and ALL of the .PREPROCESS commands
from the original netlist are commented out, even the ones that are
performed in situ (replace ground, remove redundant devices, etc.).

This commit provides a netlist that will fail to run if all the .PREPROCESS
commands are commented out.  This ensures that the bug has been fixed.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/262