Add tests for Path and Area parameter sweeps
I renamed the Gap step tests so that when I added
similar tests for the Path and Area parameteres
the filename made sense.  Now all of the model
parameters that are incorrectly scaled in the
constructor are tested with .step tests.  All of
these will fail as long as the gap, path and area
parametres are scaled in the constructor and not
in the processParams() function.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/263