Add tests for diode multiplicity
This commit adds tests for the "M=" multiplicity instance parameter
for the diode.  There is one test for transient, one for noise, and
one for sensitivity.

All three tests work basically the same way:  a netlist with two
parallel diodes is run and used as a gold standard against which the
output of a second netlist with "m=2" is compared.

The three tests cover the three different parts of the code I had to
modify to implement multiplicity correctly.

Xyce#264
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/264