Parameter step tests for Level 2 nonliner mutual inductor
The level 2 nonlinear mutual inductor model manipulates
model parameters in the constructor rather than the
processParams() method.  These added tests demonstrate
the problem and will fial as long as the level 2
model continues to modify parameters in the constructor.
The tests are duplicates of the tests for the Level 1
version of the model with file name changes and "level=2"
added as needed.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/263