Fix file name error and add offset for xyce_verify comparison
The NLAreaStepLv2.cir.sh file was using the wrong filename
for its dependent netlist.  And the current through the L2
inductor is very small and crosses zero.  The zero crossings
would lead to a lot of error in the comparison phase.  Added
and offset to all of the tests for that output.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/263