Test info sent to stdout for .FFT and .MEASURE FFT
This commit adds or updates tests that verify the info sent to
stdout if .OPTIONS FFT FFTOUT=1 is used.  It also verifies the
updated format for the <netlistName>.fftX files.  In particular:

  1) FFT/fftoutOptionZero and FFT/MiscParsing verify that the
updated stdout for .FFT is not made for .OPTIONS FFT FFTOUT=0.

  3) FFT/StartStopTest verifies the stdout for FORMAT=UNORM,
the sorting of the harmonic list, and the limiting of the values
to -200dB in the harmonic list.

  3) FFT/fft_accurateOptionZero verifies the stdout for FORMAT=UNORM.

  4) MEASURE_FFT/DefaultValTest verifies the stdout when the netlist
has both .FFT and .MEASURE lines.

  5) FFT/STEP/StdOutTest and FFT/STEP/RemeasureStdOutTest verify the
stdout for the .STEP case.  The first test's netlist has both .FFT
and .MEASURE lines.  The second test's netlist only has a .FFT line.

SON Bug 1333
Issue #186
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1333
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/186