Fix Valgrind Testing for New FFT/STEP Tests
This commit adds tags file for the two new tests for stdout in
the FFT/STEP subdirectory, since they are not part of valgrind
testing.  It also undoes the changes to the .sh file for the test
FFT/StartStopTest that were done in commit a18c438d yesterday.

SON Bug 1333
Issue #186
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1333
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/186