Test Use of Measure Names that Start with Operator Letter
This commit tests that measure names can start with an operator
letter, like V or I, if they are used as the measured quantity in
something like an EQN measure.

As part of the refactoring done to support this change, the error
messages for several other .MEASURE and .PRINT tests changed.  That
accounts for most of the other changes to various .sh files.  The
tests MEASURE/bad_dot_measure_line3 and MEASURE/bad_dot_measure_line8
were also modified to test some of the parsing corner-cases mentioned
in SON Bug 1134.

Issue #277
SON Bug 1134
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/277
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1134