Test FIND-AT Measure Behavior at First Step
This commit tests the behavior of the FIND-AT measure at the
first time, frequency, or DC step value in the analysis.
Previously, the reported measure value would be -nan for that
case.  It is now correct for all four measure modes (AC, DC,
NOISE and TRAN).  It is tested for both the step and non-step
cases for AC, DC and NOISE.  It is only tested for the non-step
case for TRAN.

Issue #283
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/283