Add tests for bsim3 soi voltage limiter flag
This is in support of this issue:
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/282
"Make it so that BSIM3 SOI voltage limiting is optionally turned off"

The two tests that I'm adding in this commit are copies of tests that
were already present in the test suite (in this directory), with the one
difference that the b3soi voltlim flag is set to 0.    I have manually
observed that the code behaves slightly differently (but still produces
a good answer) with this flag set to 0.    In the transient case, it
takes slightly fewer time steps.  So, the gold standards for the new
tests are slightly different from the originals.