Add temperature model test for soln-dep resistors
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/171
"Make Xyce directly support voltage-dependent resistors"

The resistor includes an optional temperature model which is used to
modify the nominal resistance.  Similar to the capacitor, this model
should modify the result of an resistor expression.

This test is a copy of one from the RESISTOR_TD directory, but has been
modified so that all the resistors use an expression.