Samll update to the parser tests
Related to this issue:
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/223

"Add a new subcircuit separator, "." instead of ":" as part of -hspice-ext"

This commit modifies one of the parser tests to include interface nodes
on the .PRINT line.    My work on this issue had overlooked this use
case, and until my code changes yesterday, this use case did not work.
It works now, and this test demonstrates it.

This commit also removes a spurious file that I accidentally added a
while ago.