Re-enable test
The test XDM/HSPICE/OTHER_PARSING/special_variables.cir has been
disabled for some time.

It was originally created to test certain uses of global params and
.params that was documented as illegal use but happened to work because
of a happy accident that introduced a "bug" that made it work.
Unfortunately, the bug only made SOME use cases of .params that
depended on .global_params work, and others failed horribly.

When @erkeite rewrote the expression library, he initially made it so
the documented limitation was enforced, even though the reason for it
was rapidly being removed by his rewrite.  The test was then disabled
so it wouldn't keep failing.

Over time, Eric has revamped expression and parameter handling so much
that there is no longer a need for the restriction.  The use case
tested by this test is now documented legal, and it's time to
re-enable the test.

Xyce_Regression#1
https://gitlab-ex.sandia.gov/Xyce/Xyce_Regression/-/issues/1