Add/Update DERIV Measure Tests for Issue #291
This commit adds a test (or new subtests to existing tests) for
the DERIV measure behavior at the first simulation step for AC,
DC, NOISE and TRAN measure modes.  It also adds subtests for
the case where either the measured variable or the variable in
the WHEN clause has a constant value.

Git issue #291
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/291