Align tests to how xyce_verify works
All Xyce simulations in the HSPICE category are now run with the
-hspice-ext option, including simulations of the golden netlists.
This is because xyce_verify.pl compares everything in the simulation
results, including headers. This causes problems then when comparing
results of netlists that are run with and without the -hspice-ext
flag and when the subcircuit separator character is included
in the .PRINT statement. For example, running Xyce without -hspice-ext
and printing out a node such as V(A:B) would result in an error
when using xyce_verify with a translation run with Xyce using
-hspice-ext and printing out a node V(A.B).

Since xyce_verify is used in general by Xyce Regression (and not
just the XDM portion), I did not want to touch its code. So, to
align with how it works, I changed all simulations in the HSPICE
category to use the -hspice-ext all, and the golden netlists to use
the "." separator instead. This shouldn't be a problem either, since
the XDM tests are to check that the HSPICE netlist (with extension
*.cir.hspice) are translated correctly in Xyce, and the result matches
the golden netlist's (whose result should be the same, either running
Xyce with or without the -hspice-ext flag).