Test for soln-dep resistor in subcircuits
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/171
"Make Xyce directly support voltage-dependent resistors"

It was pointed out to me today that the newly added
solution dependent resistor  will not work inside of subcircuits,
for some use cases.

Specifically, if you have a solution dependent resistor inside of a
subcircuit, and the expression depends on a node that is local to that
subcircuit, the names cannot be resolved unless the subcircuit prefix is
added to the node name.    There is code in the parser that handles
this, but it only does it for cases where solution-dependent parameters
are allowed, such as for Bsrcs, capacitors and (now) resistors.

This new test, solDepRes3 is very similar to the other tests in this
directory, except that the resistor, and the node it depends upon, are
inside of a subcircuit.