Tests for usage of coupled inductor params
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/297
"Problems using coupled inductors (from mutual inductors) on .PRINT and
.SENS lines, amongst other problems"

This commit contains several simple tests.

They are identical except for minor differences. The difference between
mutIndPrint1 and mutIndPrint2 is that the second one uses a subcircuit and
the other does not.  They produce identical numerical output.

In both of those tests, the main thing being tested is if a coupled
inductor's inductance parameter "L" can be output on the .PRINT line.

Because coupled inductors are handled in a unique manner in Xyce, the standard
parameter handling was unable to find these params for printing.  That is
fixed now so these tests work.

The 3rd test, mutIndStep, is the same as mutIndPrint1, except that it
has a .STEP statement, which sweeps the value of L1:L.  This behavior
also did not work prior to the code fixes for this issue.

This commit also includes tests for SAMPLING, EMBEDDEDSAMPLING and
intrusive PCE, using the non-expression method for specifying uncertain
parameters.