Regression test for real-inputs to modulus
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/310
"Fix modulus operator in new expression library to properly handle real
number inputs"

This test is inspired from a user netlist, in which a behavioral source
is used to produce a triangular shaped periodic signal. The source uses
(among other things) the modulus operator.  It only works correctly if
the modulus operator handles real-numbered inputs.

I have checked and this circuit produces the same answer in release 7.1,
which used the old expression library.  This type of behavioral source
was originally developed for the old expression library, so the fix to
the new expression library needed to match it.