Change Some MEASURE Tests to use Legacy Trig-Targ Mode
This commit changes the existing tests for TRAN mode TRIG-TARG
measures, in MEASURE and MEASURE/STEP, that were done for the
RiseFallDelay class to use the new (and possibly temporary)
.OPTIONS MEASURE USE_LTTM=1 setting.  This will allow these
existing tests to be continue to be run, as we transition to the
new (and more HSPICE- and ngspice-compatible) TrigTarg classes.

The test MEASURE/RiseFallDelayFT uses the default setting of
USE_LTTM=0.  This allows us to test that TRIG-TARG measures
that use the Xyce-specific FRAC_MAX qualifier will automatically
default to using the legacy RiseFallDelay class.

Gitlab issues #289, #302 and #303
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/289
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/302
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/303