Initial Set of Tests for New TrigTarg Classes
This is the initial set of tests for the new TrigTarg classes.
They cover AC and NOISE measure modes for both the non-step and
.STEP cases.  A non-step test (with an ascending DC sweep only)
was added for the DC measure mode.

Additional tests (or subtests) that are needed are:

  1) Tests for AC_CONT, DC_CONT, NOISE_CONT and TRAN_CONT modes.

  2) Tests for TRAN mode.

  3) Subtests in MEASURE_AC for S-params and operators beside VM().

  4) Subtests in MEASURE_NOISE for the various noise operators,
and operators other than VM().

  5) Subtests for more complicated TRIG or TARG clauses such as
TRIG VM(a)=VM(b), and also expression support.

  6) Subtests for failed measures.

Issues #220, #289, #302 and #303
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/220
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/289
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/302
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/303