More Tests for TRIG-TARG for Continuous Modes
This commit adds additional tests for TRIG-TARG measures for the
AC_CONT, NOISE_CONT and TRAN_CONT measure modes.  It covers both
the non-step and .STEP cases for those modes, although the TRAN_CONT
tests still need to test the RISE and FALL qualifiers.

A TRIG-TARG measure was added to the various tests in the
MEASURE_CONT/USE_CONT_FILES subdirectory to make sure that the
results for continuous mode TRIG-TARG measures are sent to the
correct filenames.

Issues #289, #302 and #303
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/289
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/302
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/303