Make it possible to run plugin tests on CEE
The VERILOGPLUGIN tests were written to make it possible to test
Verilog plugins under Jenkins, and as such were really never made
properly flexible.  They ran under those jobs on CEE, but the
"installed" test installs in a local directory and provides a full
path to the binary.

I just tried running a little script of mine to test the *actual*
installed binaries on CEE that are accessed through the module, and
discovered that the test scripts in VERILOGPLUGIN were too stupid.
They try to strip out the path of the Xyce binary and sanity check for
the presence of buildxyceplugin and that the lib directory contains
shared libraries, but that doesn't work if the "path" to Xyce is just
"Xyce".  It also doesn't work if the executable name is, say
"/usr/local/mpi/openpi/bin/mpirun -np 2 /path/to/Xyce" or "mpirun -np
2 Xyce".

That is, it doesn't work if Xyce is in your path and you just call it
"Xyce" or if we're running in parallel.  This commit fixes that.

It also adds parallel to the tags list.

Now I am able to test the actual, for real installed Xyce on CEE using
the exact module that customers would use, *AND* it includes testing
of verilog-a plugins.

Had I had these tests in place and working prior to the 7.3 release
then I'd have detected Xyce#260 during QA-ON-COE stage instead of as
the result of having a customer's plugin puke.