Initial Set for Tests for TRAN Mode TRIG-TARG Measure
This is the initial set of tests for the new TrigTarg classes
for the TRIG-TARG measure for TRAN measure mode.  They cover
these subtests for both the non-step and STEP cases:

  1) TRIG-AT or TARG-AT, or both TRIG-AT and TARG-AT.

  2) Various values for CROSS in both TRIG and TARG, including
CROSS=LAST and CROSS=-1.

  3) Negative values for the measure.

  4) A TD value for TRIG, TARG or both.

This commit also added a subtest to MEASURE_AC/TrigTargTestAC
to verify that branch currents work, and that both the real
and imaginary parts of the solution vector are accessible.

Issues #220 and #289
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/220
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/289