Add nonlinear coupled inductor sensitivity tests
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/297
"Problems using coupled inductors (from mutual inductors)
on .PRINT and .SENS lines, amongst other problems"

This commit adds 2 tests for sensitivity analysis, where the sensitivity
parameter is one of the coupled inductor values, L1:L.

In both cases, they are compared to a netlist-level finite difference.
In other words, a netlist in which the circuit is copied, but with a
slightly different value for L1:L.  The computed sensitivity from
the .SENS netlist is then compared to a finite difference expression
from the FD netlist.

I'll note that for the level=2 sensitivity calculation, in my initial
attempt at it, the resulting sensitivity had sawtooth oscillations.  I
made those go away by setting the time integration method to gear, and
tightening the tolerances.