Test for fix to .PREPROCESS REPLACEGROUND for .SUBCKT, etc.
This test starts with the SUBCKT/subckt_a1.cir and adds the
option to replace ground node synonyms with "0".  To this netlist
a change has been made to move the subcircuit definitions into an
include file named "GND".  This ensures that if the ground node
preprocessing is enabled, it does not process and replace ground
synonyms in include or library file names or definitions, as well
as within subcircuit definitions.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/318