Update MIXED_SIGNAL Tests for commit Xyce@6583c731
Commit Xyce@6583c731 removed some debugging output from the
file utils/XyceCInterface/N_CIR_XyceCInterface.C, per a request
from an external collaborator.  During the "pre-historic phase" of
getting the Python form of the Mixed Signal Interface to work, it
was deemed useful to check for that string in the test output.  That
is no longer useful, so the relevant .sh files in MIXED_SIGNAL/Python
are also being updated.