Finish up Tests for TRIG-TARG Measures for DC and DC_CONT
This should mostly finish up testing for TRIG-TARG measures, that use
the new TrigTarg classes, for DC and DC_CONT measure modes.  It adds
subtests for the .STEP and "non-step" cases for failed measures,
and the TD, RISE and FALL qualifiers.  The CROSS=LAST and CROSS=-1
values are tested for DC measure mode only.  Issue #289 will discuss
why those features are not tested for DC_CONT mode.

Also, a test for DC_CONT mode for the non-step case was added.
The other new test (MEASURE_DC/TrigTargTest2DC) covers the case
of two sweep variables on the .DC line rather one being on the
.STEP line.  I'll add commentary on that test case to Issue #289.

Issues #289 and #302
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/289
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/302