Additional AT tests for TRIG-TARG measures
This commit includes additional .MEASURE lines in the Trig-Targ
netlists for the MEASURE_AC and MEASURE_NOISE test cases. The
additional lines are intended to test the use of AT in the TRIG clause
and in the TARG clause (a supported feature in the new TrigTarg class).

These changes are part of the subtests outlined under Item (4) of note
95559 on this issue.
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/289#note_95559

The changes to the gold standard files were verified in hspice.

Issue #289
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/289