Update for handling of unrecognized options
This is related to this issue:
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/237
"add parser support for ".option scale" so that ".options parser scale"
doesn't have to be translated via XDM"

I recently addressed this issue in the Xyce source code.  One side
effect is that some malformed .OPTIONS statements, which were previously
treated as fatal errors are now treated as warnings, and the output
message is slightly different.  That made it necessary to update this
test.