Add tests for option scale
This is in support of this issue:
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/237
"add parser support for ".option scale" so that
".options parser scale" doesn't have to be translated via XDM"

This commit adds 4 tests, which are copies of previously developed tests
for "options parser scale".   The only difference with these is that
the scale parameter is specified as ".option scale=#" instead of ".options
parser scale=#".