Update Tests for Change in MEASFAIL Option Behavior
This commit updates various .MEASURE tests for the change in how
Xyce reports failed measures in the output files.  The tests in
MEASURE_AC, MEASURE_DC, MEASURE_CONT, MEASURE_FFT, and MEASURE_NOISE
now all use the default of .OPTIONS MEASURE MEASFAIL=1.  All of the
tests in those directories now output "FAILED" to the measure files
for failed measures.  The default value for a failed measure is zero.
In many  cases, the failed measures were renamed since the old name
no longer reflected what the subtest was now testing.

The MEASURE and MEASURE/STEP directories are where various permutations
of .OPTIONS MEASURE MEASFAIL=<0|1> are then tested.  The tests that use
the "old style of verification" in MEASURE of calculating the measure
value in a Perl script in the test's .sh file generally use MEASFAIL=0,
and test the default calculation of 0 and often also explicitly setting
another default value (typically -100).  The tests in MEASURE/STEP
mostly use MEASFAIL=0 and DEFAULT_VAL=-1.  This verifies that the
default calculation value is reset for each .STEP iteration.  It
also verifies that the Xyce 7.4 behavior can be reproduced.

MEASURE/DefaultValOption verifies that .OPTIONS MEASURE DEFAULT_VAL=<val>
takes precedence over the DEFAULT_VAL qualifier on the .MEASURE lines
when .OPTIONS MEASURE MEASFAIL=0 is used.  MEASURE/MeasfailOption
verifies the use of  .OPTIONS MEASURE MEASFAIL=0  when the DEFAULT_VAL
qualifier is used on .MEASURE lines.  This may be duplicative with
other tests, but it does make sure to test that functionality for all
measure types.

Issue #304
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/304