Add test for units of C
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/328
"Support "C" for temperature units"

While working on a different bug (for option scale), I discovered that
there was one test in the test suite (BUG_1788)  which specified
temperature as:

.option TEMP=27C

My work on "option scale" made it so that this line

.option scale=1.0e-6

 would be auto-converted to

.options device scale=1.0e-6

during parsing.  That had the side effect of auto-converting the
".option temp=27C" into "options device temp=27C".  There
was already old code to this effect in Xyce, but my work made it actually
get executed in the BUG_1788 test.

So far, so good, except that Xyce did not like the "C" unit.  It
intepretted "27C" as an expression instead of a numerical value, and
then exited with an error. (since most options parameters are not
allowed to be expressions, with a handful of exceptions)

Fixing this was trivial, and so I did that last week.  The "isValue"
function in the N_UTL_ExtendedString.C file has a list of allowed units.
So, all I had to do was add "C" to that list.

I initially thought that the existing test which used the above line was
sufficient, but today had second thoughts.  The existing test doesn't
really check if the temperature is interpretted correctly.  It
just ensures that the code doesn't exit with an error.

This commit adds a new, very simple test.  I took one of the bsim4
tests, which tests the temperature to 100C, and then created several
versions of this netlist.  The 3 versions specify temperature in the
following ways:

bsim4_temp.cir:   .options device Temp=100.0
bsim4_tempC.cir:  .options device Temp=100.0C
bsim4_tempC2.cir: .option Temp=100.0C

Then the script checks to make sure they all produce the same answer.

This commit also adds the "issue328" tag to the BUG_1788 test.