Removing unnecessary options from test
The NLNEARCONV option used in both these tests is no longer needed for
the test to successfully complete.  Given that this is a non-ideal option
that enables the nonlinear solver to declare convergence when it has not,
it is best not to use it when it is not necessary.

This was found as part of the nonlinear solver work in issue:

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/329