Update ADC tests for dynamic state history arrays
This test calls Simulator::getTimeStatePairs() through python and
the XyceCInterface class.  Prior to my commit
d677f883a92274c5efa5baa914f6a76ce602de5c calling getTimeStatePairs()
had the side effect of clearing the time-voltage history in the ADC.
The operator class TimeStatePairsOp always adds in the current time
and state, so the history vectors always had something in them.
That let this test assume that the latest time points were always
at the base of the history array.  Also, different ADC's will have
different history lengths depending on their width and range.
This test uses two ADC's of differnt width that change state at
different times in the simulation, but it assumed that the history
arrays will be the same length.  That issue has been corrected.
Finally, the origina "gold" standard data was reporting the history
wrong so that the output values would jump in time from the current
simulation time to zero and back.  This was not consistent with
the state as it is output in the simulation's PRN file.  Now one
can look at the PRN and verify that the reported state through
getTimeStatePairs() is the SAME as what was output in the
PRN at that time.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/338