Fix VPI tests for dynamic state arrays
In the same way that my previous commit fixed
the python test case to allow for dynamic (i.e.
changing in length during the simulation) state
arrays, this commit fixes the VPI tests.  This
test assumed that the ADC state arrays were
always two elements long and that all ADC's
had the same length history.  To fix this
I had to change the logic in the test
program and update the gold standard as it
wasn't outputting the ADC state correct
for Xyce's actual simulation time.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/338