Fixes netlists to use a non-default maxstep or better gold standard
When the logic in the DampedNewton solver is corrected to
only take the maximum number of steps, and not one more, then
these tests fail due to convergence issues or comparison.
Setting the maxstep to 21 returns the behavior of the test to
before the correction was made in the DampedNewton logic.
In the case of the atanh_tanh test, the comparison offsets were
wrong and the gold standard was too loose.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/329