Exit correctly from script if Xyce errors out
This script carefully checks if either of the two netlists it runs has
failed, printing an error message to "$cirfile.err" and setting a
flag.

Unfortunately, the file it prints two is not one of those that gets
shipped off to CDash if the test fails, and nothing downstream of that
actually checks the status of the flag --- it is presumed that Xyce
has produced both output files and the test tries to verify them.

If either netlist had failed, the result would be reported as a
compare failure, when in fact it should have been an "EXITED WITH
ERROR".

That was my mistake when I created this test script.  Mea culpa, mea
culpa, mea maxima culpa.

This commit adds appropriate error checking.

I discovered this problem when trying to diagnose mysterious MPI test
failures on OS X with more than 2 processors.

In addition to making sure we exit with the correct error code, I am
adding some code to copy the output of the failing netlist out of the
file where it is dumped and into a file which is actually uploaded to
CDash.  We only upload "$CIRFILE.out" and "$CIRFILE.err" on failure,
where $CIRFILE is the name of the test.  Many of our test cases run a
bunch of netlists with names OTHER than $CIRFILE, and redirect the
output to someplace other than $CIRFILE.out or $CIRFILE.err.  This
makes it really hard to diagnose problems when they fail.