Correct handling of return time, value/state data
The ADC data from getTimeVoltagePairs and getTimeStatePairs
was not consistently checked so that the right number
of values were output.  Under some conditions there is only
one point per ADC and othertimes mulitple points.  It depends
on if the given ADC has had a state change during the last
call to simulateUntil.  Now the python and VPI tests cases
check when dealing with ADC output. This is consistent
with the examples in the mixed signal application note.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/338