Fix truncated output in test and gold data
The test program ADCtest.c was using a "%d" format
specifier for a double type output.  That produced
truncated output with icc and very large output
with gcc.  Corrected the format specifier to be a
"%.3e" to match the time data.  Updated the
gold standard as well since it had lots of
truncated data.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/338