Test using xyce_getTimeVoltagePairsADCLimitData()
This is a python based test that uses the C function
xyce_getTimeVoltagePairsADCLimitData() in XyceCInterface.  It
specifically uses this function to get the ADC data of
an pair of ADC's that will have different numbers of values
in the history array during the simulation. This has also
been tested to work with python3 as well.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/348