Update to work with python3
Updated the scripts to work with both Python2.7 and Python3.  The
actual python used will depend on what the test system maps
to the command "python".  To check that the worked with python 3
I manually changed all the perl calls to system("python ...")
to system("python3 ...").  That worked on my local mac but this will
need a test system where "python" is indeed python 3.x.  Also, I
only verified that tests in this directory worked with Python 3.
There may be other tests in Xyce_Regression that do not work with
Python 3.  That will have to be another issue.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/351