Update expected (gold) output for .SENS test cases
The output from .print sens lines was formatted slightly differently
across the different analysis types (different column header
formatting - gitlab issue 270). The format for the column headers
should be consistent. This commit updates the headers in the gold
standard output files for the sensitivity analysis test cases.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/270

Remove property name editing from convert scripts

There are several test cases for verifying the different output
formats from Xyce. Instead of maintaining a gold standard output file
for each format, some test cases convert the test output and compare
to another standard output file. This is mostly done in the test cases
for the tecplot output. The changes in this commit are to update the
converter scripts to account for some formatting changes associated
with issue 270. The column header format was modified for sensitivity
analysis data. This change is to keep the field names consistent
through the conversion process.

Update test cases using convertXyceFormat script

This commit updates the convertXyceFormat.py scripts to account for the
column header formatting for sensitivity analysis data. When converting
from one format to another, the script was removing curly braces and
spaces from the field names. Leading and trailing whitespace is still
removed, but the data field names/format should now remain consistent
in non-standard format output files are converted to the standard PRN
format.

Remove ac sens test for acobjfunc and objvars input

With changes for gitlab issue 270, objvars parameters are processed by
the same functions as acobjfunc parameters (basically a lot of the
objvars handling was removed while maintaining backwards compatibility).
As a result of using common code to handle both forms for declaring
objective functions, the restriction on using both on the same .SENS
line in a netlist is no longer needed (and the check against this was
removed from Xyce). This commit modifies the test that verified the old
check; it has been repurposed to check the error condition when no
parameters for the objective function(s) (i.e. missing PARAM=... on a
.SENS line).